c.lib 1.174 (27.8.2004)
- The parameters of atan2() were swapped. Fixed.

- Merged fdlibm 5.3 changes with __kernel_tan(), __exp() and __pow()
  functions.

+++ math_exp.c
@@ -194,7 +194,7 @@ __exp(double x)
 	    } 
 		else 
 		{
-			k  = invln2*x+halF[xsb];
+			k  = (int)(invln2*x+halF[xsb]);
 			t  = k;
 			hi = x - t*ln2HI[0];	/* t*ln2HI is exact here */
 			lo = t*ln2LO[0];
